Updated README.md, fixed typo in SQL file, fixed bug with viewing books
Before this commit, if two books with the same title were added, when trying to view those books, the ResultSet would always view the first row of the book no matter what.
The code now accounts for multiple books of the same title so that we can properly display the appropriate attributes.